Allow setting mutliple fonts
Also fix icon alignment
#970

+++ trview.app/Settings/FontSetting.h
@@ -8,6 +8,6 @@ namespace trview
     {
         std::string name;
         std::string filename;
-        float       size;
+        int         size;
     };
 }
